requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.4.0
python-dotenv>=1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
alembic>=1.13.0

# Optional (development/testing)
pytest>=7.0.0